+++ backend/swagger/docs/docs.go
@@ -2653,11 +2653,11 @@ const docTemplate = `{
                     "type": "boolean"
                 },
                 "externalUrl": {
-                    "description": "used by share links if set",
+                    "description": "used by share links if set (eg. http://mydomain.com)",
                     "type": "string"
                 },
                 "internalUrl": {
-                    "description": "used by integrations if set, this is the url that an integration service will use to communicate with filebrowser",
+                    "description": "used by integrations if set, this is the base domain that an integration service will use to communicate with filebrowser (eg. http://localhost:8080)",
                     "type": "string"
                 },
                 "logging": {
@@ -2908,6 +2908,10 @@ const docTemplate = `{
                 },
                 "showHidden": {
                     "description": "show hidden files in the UI. On windows this includes files starting with a dot and windows hidden files",
+                    "type": "boolean"
+                },
+                "showSelectMultiple": {
+                    "description": "show select multiple files on desktop",
                     "type": "boolean"
                 },
                 "singleClick": {
@@ -3206,6 +3210,10 @@ const docTemplate = `{
                 },
                 "showHidden": {
                     "description": "show hidden files in the UI. On windows this includes files starting with a dot and windows hidden files",
+                    "type": "boolean"
+                },
+                "showSelectMultiple": {
+                    "description": "show select multiple files on desktop",
                     "type": "boolean"
                 },
                 "singleClick": {